users/models.py
from django.db import models
from django.contrib.auth.models import User
from PIL import Image
from reservations.models import Reservations

class Profile(models.Model):
    user = models.OneToOneField(User, on_delete = models.CASCADE)
    first_name = models.CharField(max_length=30, blank=False)
    last_name = models.CharField(max_length=30, blank=False)
    phone_number = models.IntegerField(blank=True, null=True)
    image = models.ImageField (default = 'default.png', upload_to='profile_pics')


    def __str__(self):
        return self.user.username

    def save(self, *args, **kwargs):
        super().save(*args, **kwargs)

        img= Image.open(self.image.path)

        if img.height > 300 or img.width > 300:
            output_size = (300,300)
            img.thumbnail(output_size)
            img.save(self.image.path)
