requirements.txt
boto3==1.9.165
botocore==1.12.165
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
Pillow==6.0.0
psycopg2-binary==2.8.2
pypi==2.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
stripe==2.35.1
urllib3==1.25.3
virtualenv==16.6.0
whitenoise==4.1.2
